create util for asset list creation
