+++ packages/client/hmi-client/src/page/project/components/tera-project-overview.vue
@@ -165,8 +165,8 @@
 </template>
 
 <script setup lang="ts">
-import { isProjectAssetTypes } from '@/types/Project';
 import { computed, nextTick, onMounted, ref } from 'vue';
+import { generateProjectAssetsList } from '@/utils/project-assets-list';
 import InputText from 'primevue/inputtext';
 import Button from 'primevue/button';
 import Menu from 'primevue/menu';
@@ -175,8 +175,7 @@ import Column from 'primevue/column';
 import * as DateUtils from '@/utils/date';
 import TeraAsset from '@/components/asset/tera-asset.vue';
 import CompareModelsIcon from '@/assets/svg/icons/compare-models.svg?component';
-import { Tab } from '@/types/common';
-import { capitalize, isEmpty } from 'lodash';
+import { capitalize } from 'lodash';
 import { AssetType } from '@/types/Types';
 import { useRouter } from 'vue-router';
 import { RouteName } from '@/router/routes';
@@ -196,42 +195,11 @@ const openedRow = ref(null);
 const searchTable = ref('');
 const showMultiSelect = ref<boolean>(false);
 
-const assets = computed(() => {
-	const tabs = new Map<AssetType, Set<Tab>>();
-
-	const projectAssets = useProjects().activeProject.value?.assets;
-	if (!projectAssets) return tabs;
-
-	const result = <any>[];
-	// Run through all the assets type within the project
-	Object.keys(projectAssets).forEach((type) => {
-		if (isProjectAssetTypes(type) && !isEmpty(projectAssets[type])) {
-			const projectAssetType: AssetType = type as AssetType;
-			const typeAssets = projectAssets[projectAssetType]
-				.map((asset) => {
-					let assetName = (asset?.name || asset?.title || asset?.id)?.toString();
-
-					// FIXME should unify upstream via a summary endpoint
-					if (asset.header && asset.header.name) {
-						assetName = asset.header.name;
-					}
-
-					const pageType = asset?.type ?? projectAssetType;
-					const assetId = asset?.id ?? '';
-					return { assetName, pageType, assetId };
-				})
-				.filter((asset) => {
-					if (!searchTable.value?.trim()) {
-						return true;
-					}
-					const searchTermLower = searchTable.value?.trim().toLowerCase();
-					return asset.assetName.toLowerCase().includes(searchTermLower);
-				});
-			result.push(...typeAssets);
-		}
-	});
-	return result;
-});
+const assets = computed(() =>
+	Array.from(generateProjectAssetsList(searchTable.value).values())
+		.map((set) => Array.from(set))
+		.flat()
+);
 
 const onRowSelect = (selectedRows) => {
 	// show multi select modal when there are selectedRows otherwise hide
